+++ lib/hnco/algorithms/ea/mu-comma-lambda-ea.cc
@@ -46,7 +46,7 @@ MuCommaLambdaEa::init()
 void
 MuCommaLambdaEa::iterate()
 {
-  for (size_t i = 0; i < _offsprings.size(); i++) {
+  for (int i = 0; i < _offsprings.size(); i++) {
     bit_vector_t& offspring = _offsprings.get_bv(i);
     offspring = _parents.get_bv(_select_parent(Random::generator));
     _mutation.mutate(offspring);
