+++ lib/hnco/algorithms/hea/hea.hh
@@ -140,7 +140,7 @@ namespace hea {
           _herding->init();
       }
 
-      for (size_t i = 0; i < _population.size(); i++)
+      for (int i = 0; i < _population.size(); i++)
         _herding->sample(_target, _population.get_bv(i));
 
       if (_log_flags[LOG_ERROR])
