+++ lib/hnco/algorithms/pv/nps-pbil.cc
@@ -36,7 +36,7 @@ NpsPbil::init()
 void
 NpsPbil::iterate()
 {
-  for (size_t i = 0; i < _population.size(); i++)
+  for (int i = 0; i < _population.size(); i++)
     pv_sample(_pv, _population.get_bv(i));
 
   if (_functions.size() > 1)
